% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.assess_sscores_2.R
\name{assess_score}
\alias{assess_score}
\title{Assess PRS prediction accuracy}
\usage{
assess_score(
  trait,
  plink_prefix,
  pgslist,
  phenofile,
  ancestry,
  pheno_name,
  isbinary,
  out,
  step1 = F
)
}
\arguments{
\item{trait}{Trait name}

\item{plink_prefix}{The genotype file in plink format (bed/bim/fam)}

\item{pgslist}{PGS list to evaluate}

\item{phenofile}{Phenotype file}

\item{ancestry}{Ancestry}

\item{pheno_name}{Phenotype name in the phenotype file}

\item{isbinary}{Is this trait binary?}

\item{out}{Output prefix}

\item{step1}{is this running step 1 to get accuracy of the null model}
}
\value{
The dataframe of prediction accuracy of all scores
}
\description{
This function estimates PRS prediction accuracy for a list of PRS
}
