% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.6.combine_PGS_1.R
\name{combine_PGS}
\alias{combine_PGS}
\title{Combine PGS}
\usage{
combine_PGS(
  trait = "cad",
  anc = "eur",
  pgslist = "cad_list.txt",
  pheno_name = "CAD_case",
  isbinary = T,
  phenofile = "/home/jupyter/data/phenotypes/CAD_revised.csv",
  out = "test_cad_mixedPRS"
)
}
\arguments{
\item{trait}{The name of the trait}

\item{anc}{Intended ancestry}

\item{pgslist}{PGS list of the trait}

\item{pheno_name}{Name of the phenotype column}

\item{isbinary}{True if this is binary}

\item{phenofile}{Directory to the phenotype file}

\item{out}{Output prefix}
}
\value{
Prediction accuracy of PRSmix
}
\description{
This function combines multiple PGS into one score
}
