% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_PRS.R
\name{combine_PRS}
\alias{combine_PRS}
\title{Perform linear combination of the polygenic risk scores}
\usage{
combine_PRS(
  pheno_file,
  covariate_file,
  score_files_list,
  trait_specific_score_file,
  pheno_name,
  isbinary,
  out,
  allPGS_list = NULL,
  metascore = NULL,
  liabilityR2 = F,
  IID_pheno = "IID",
  covar_list = c("age", "sex", paste0("PC", 1:10)),
  cat_covar_list = NULL,
  ncores = 1,
  is_extract_adjSNPeff = F,
  original_beta_files_list = NULL,
  train_size_list = NULL,
  train_ids_file = NULL,
  training_result_file = NULL,
  power_thres_list = c(0.95),
  pval_thres_list = c(0.05),
  nfold_cv = 3,
  read_pred_training = FALSE,
  read_pred_testing = FALSE,
  debug = F
)
}
\arguments{
\item{pheno_file}{Directory to the phenotype file}

\item{covariate_file}{Directory to file with covariate information (age,sex,PC1..10)}

\item{score_files_list}{A vector contain directories of the PGS to be read}

\item{trait_specific_score_file}{A filename contain PGS IDs of trait-specific to combine (PRSmix), one score per line}

\item{pheno_name}{Column name of the phenotype in pheno_file}

\item{isbinary}{True if this is binary}

\item{out}{Prefix of output}

\item{metascore}{Meta-information from PGS Catalog contain PGS id and trait names. Must contains information for ALL the scores (DEFAULT = NULL)}

\item{liabilityR2}{TRUE if liability R2 should be reported (DEFAULT = FALSE)}

\item{IID_pheno}{Column name of IID of phenotype file (e.g IID, person_id)}

\item{covar_list}{A vector of of covariates, must exists as columns in covariate_file (DEFAULT = age, sex, PC1..10))}

\item{ncores}{Number of CPU cores for parallel processing (DEFAULT = 1)}

\item{is_extract_adjSNPeff}{TRUE if extract adjSNPeff, FALSE if only calculate the combined PRS. May consume extended memory (DEFAULT = FALSE)}

\item{original_beta_files_list}{The vector of directories to SNP effect sizes used to compute original PRSs (DEFAULT = FALSE)}

\item{train_size_list}{A vector of training sample sizes. If NULL, all 80\% of the samples will be used (DEFAULT = NULL)}

\item{train_ids_file}{A file containing the iids to train}

\item{power_thres_list}{A vector of power thresholds to select scores (DEFAULT = 0.95)}

\item{pval_thres_list}{A vector of P-value thresholds to select scores (DEFAULT = 0.05)}

\item{read_pred_training}{TRUE if PRSs were assessed in the training set was already run and can be read from file (DEFAULT = FALSE)}

\item{read_pred_testing}{TRUE if PRSs were assessed in the testing set was already run and can be read from file (DEFAULT = FALSE)}
}
\value{
This function will return several files including
\itemize{
\item The case counts (for binary trait),
\item The dataframe of training and testing sample split from the main dataframe,
\item The prediction accuracy for each PRS in the training and testing set,
\item The prediction accuracy assessed in the testing set of the best PRS selected from the training set,
\item the AUC of the NULL model of only covariates, the best PGS, PRSmix and PRSmix+ (adjusted for covariates),
\item Odds Ratio of the best PGS, PRSmix and PRSmix+ (adjusted for covariates),
\item The mixing weights of the scores used in combination,
\item The adjusted SNP effects to estimate PRSmix and PRSmix+ (if is_extract_adjSNPeff=TRUE)
\item Return 0 if no error
}
}
\description{
This function perform a linear combination of the scores
}
