% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_PRS.R
\name{estimate_PRS}
\alias{estimate_PRS}
\title{Estimate PRS}
\usage{
estimate_PRS(plink_exe, geno, ref_file, out, start_col = 4)
}
\arguments{
\item{geno}{Genotype file in plink format (bed/bim/fam).}

\item{ref_file}{The reference risk-increasing allele file.}

\item{out}{Name of output file.}

\item{start_col}{Index of starting column to estimate PRS in the reference file (default = 4).}

\item{wfile}{The weight file.}
}
\value{
Execute plink. Return 0 if successful.
}
\description{
This function estimate PRS with a reference scores
}
