% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PRS_acc.R
\name{eval_multiple_PRS}
\alias{eval_multiple_PRS}
\title{Evaluate multiple PRSs from the given vector of PRS}
\usage{
eval_multiple_PRS(
  data_df,
  pgs_list,
  covar_list,
  liabilityR2 = F,
  alpha = 0.05,
  isbinary = F,
  ncores = 1,
  regression_output = F,
  pheno = "trait"
)
}
\arguments{
\item{data_df}{A dataframe where rows for samples, columns for PRS}

\item{pgs_list}{PGS list to evaluate, must be exist as columns in the dataframe}

\item{covar_list}{Array of covariates, must be exist as columns in the dataframe}

\item{liabilityR2}{TRUE if liability R2 should be reported (default = FALSE)}

\item{alpha}{Significance level to estimate power (default = 0.05)}

\item{isbinary}{TRUE if binary and FALSE otherwise (default = FALSE)}
}
\value{
A dataframe for prediction accuracy of multiple PRSss and their power with R2, R2 for output form, standard error, lower 95\% CI, upper 95\% CI, P-value and power
}
\description{
This function get prediction accuracy of the list of PRS
}
