% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varPRS_theory.R
\name{eval_prs}
\alias{eval_prs}
\title{Evaluate PRS prediction accuracy}
\usage{
eval_prs(
  data_df,
  prs_name,
  covar_list,
  isbinary = F,
  liabilityR2 = F,
  alpha = 0.05,
  debug = F
)
}
\arguments{
\item{data_df}{Data to assess prediction accuracy}

\item{prs_name}{PGS list of the trait}

\item{covar_list}{Array of covariates}

\item{isbinary}{TRUE if binary and FALSE otherwise}

\item{debug}{TRUE to verbose debugging}
}
\value{
A dataframe for prediction accuracy of PRS and their power
}
\description{
This function get R2 for PRS accuracy
}
