% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_adjSNPeff.R
\name{extract_adjSNPeff}
\alias{extract_adjSNPeff}
\title{Combine PGS}
\usage{
extract_adjSNPeff(mixing_weight_file, snp_eff_files_list, out)
}
\arguments{
\item{mixing_weight_file}{The mixing weight files, 2 columns: 1) Score ID 2) Mixing weights}

\item{snp_eff_files_list}{A vector of names of file of original SNP weights}

\item{out}{Name of the output file}
}
\value{
Adjusted SNP effects
}
\description{
This function combines multiple PGS into one score
}
