% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_PGSid.R
\name{extract_PGSid}
\alias{extract_PGSid}
\title{Get PGS IDs from PGS Catalog using EFO id}
\usage{
extract_PGSid(data, trait, efo)
}
\arguments{
\item{data}{directory to metadata score from PGS catalog (pgs_all_metadata_scores.csv)}

\item{trait}{The name of the trait}

\item{efo}{Ontology Trait ID (EFO from PGS catalog)}
}
\value{
a dataframe of PGS id from PGS catalog and a file with the PGS ID on each row
}
\description{
This function extract all PGS IDs (e.g. PGS000013, ...) from EFO code (e.g. EFO_0001645) from PGS Catalog
}
