% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.1.harmonize_risk_allele.R
\name{get_risk_allele}
\alias{get_risk_allele}
\title{Get risk allele by consensus across multiple PRS snp effect panels}
\usage{
get_risk_allele(
  traitname = "cad",
  pgslist = "~/data/optimization/cad/cad_list.txt",
  plinkfile = "AoU_98K_WGS_QCed_callrate.0.9_hwe.1e-15_maf0.0001_eur",
  anc = "eur",
  out = "eval_cad_pgs.txt"
)
}
\arguments{
\item{traitname}{Name of the trait}

\item{pgslist}{PGS list of the trait}

\item{plinkfile}{Plink file}

\item{anc}{Ancestry}

\item{out}{Output}
}
\value{
A dataframe of SNP and risk-increasing allele
}
\description{
This function get the risk-increasing allele based on consensus across PRS panels
}
