% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.5.2.harmonize_risk_allele.R
\name{harmonize_snpeffect}
\alias{harmonize_snpeffect}
\title{Harmonize snp effect with provided risk allele}
\usage{
harmonize_snpeffect(wfile, ref, out)
}
\arguments{
\item{wfile}{The weight file}

\item{ref}{The reference risk-increasing allele file}

\item{out}{Name of output file}
}
\value{
A dataframe of SNP and risk-increasing allele
}
\description{
This function harmonize SNP effect according the panel of risk allele
}
