% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.5.harmonize_spneffect_alltraits.R
\name{harmonize_snpeffect_toALT}
\alias{harmonize_snpeffect_toALT}
\title{Harmonize snp effects across PGS to the alternative allele}
\usage{
harmonize_snpeffect_toALT(freq_file, pgs_folder, pgs_list, out)
}
\arguments{
\item{freq_file}{Frequency file with ID, ALT, REF columns (e.g from PLINK2)}

\item{pgs_folder}{Directory to folder contain PGS}

\item{pgs_list}{File contain PGS id on each line}

\item{out}{Filename of the output}
}
\value{
0 if no error
}
\description{
This function harmonize SNP effect across PGS to the alternative allele
}
