% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_columns.R
\name{make_dummy_columns}
\alias{make_dummy_columns}
\title{Make dummy columns for categorical variables}
\usage{
make_dummy_columns(df, cat_covar_list)
}
\arguments{
\item{df}{Dataframe with variables}

\item{cat_covar_list}{A vector of categorical variables}
}
\value{
This function will generate a list containing a dataframe and its dummy columns for categorical variables
\itemize{
\item Return a list containing a dataframe and its dummy columns for categorical variables
}
}
\description{
This function makes dummy columns for categorical variables
}
