% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.6.combine_PGS_1.R
\name{combine_PGS}
\alias{combine_PGS}
\title{Combine PGS}
\usage{
combine_PGS(
  trait = "cad",
  anc = "eur",
  pgslist = "cad_list.txt",
  pheno_name = "CAD_case",
  isbinary = T,
  score_files_list = "",
  basic_data_file = "/home/jupyter/data/phenotypes/aou_basic_data.csv",
  metascore = "~/data/pgs_all_metadata_scores.csv",
  phenofile = "/home/jupyter/data/phenotypes/CAD_revised.csv",
  score_pref = "AoU_98K_WGS_QCed_callrate.0.9_hwe.1e-15_maf0.0001_",
  out = "test_cad_mixedPRS_eur",
  read_pred_training = F,
  read_pred_testing = F
)
}
\arguments{
\item{trait}{The name of the trait}

\item{anc}{Intended ancestry}

\item{pgslist}{PGS list of the trait}

\item{pheno_name}{Name of the phenotype column}

\item{isbinary}{True if this is binary}

\item{basic_data_file}{Directory to covariate information}

\item{phenofile}{Directory to the phenotype file}

\item{score_pref}{Prefix of score files}

\item{out}{Output prefix}

\item{read_pred_training}{True if the training set PRS assessment already run and can be read from file}
}
\value{
Prediction accuracy of PRSmix
}
\description{
This function combines multiple PGS into one score
}
