% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.6.combine_PGS_1.R
\name{combine_PGS}
\alias{combine_PGS}
\title{Combine PGS}
\usage{
combine_PGS(
  trait,
  anc,
  pgslist,
  pheno_name,
  isbinary,
  score_files_list,
  basic_data_file,
  metascore,
  phenofile,
  score_pref,
  out,
  liabilityR2 = F,
  IID_pheno = "person_id",
  covar_list = c("age", "sex", paste0("PC", 1:10)),
  ncores = 5,
  train_size_list = NULL,
  power_thres_list = c(0.95),
  pval_thres_list = c(0.05),
  read_pred_training = NULL,
  read_pred_testing = NULL
)
}
\arguments{
\item{trait}{The name of the trait}

\item{anc}{Intended ancestry}

\item{pgslist}{PGS list of the trait}

\item{pheno_name}{Name of the phenotype column}

\item{isbinary}{True if this is binary}

\item{score_files_list}{A list contain PGS directory}

\item{basic_data_file}{Directory to file with covariate information (age,sex,PC1...,PC10)}

\item{metascore}{Meta-information contain PGS id and trait names}

\item{phenofile}{Directory to the phenotype file}

\item{score_pref}{Prefix of score files}

\item{out}{Output prefix}

\item{IID_pheno}{Column name of IID of phenotype file (e.g IID, person_id)}

\item{covar_list}{Array of covariates}

\item{ncores}{Number of cores to run}

\item{power_thres_list}{Power threshold to select scores}

\item{read_pred_training}{True if the training set PRS assessment already run and can be read from file (Default: FALSE)}

\item{read_pred_testing}{True if the training set PRS assessment already run and can be read from file (Default: FALSE)}

\item{pval_thres}{P-value threshold to select scores}
}
\value{
Prediction accuracy of PRSmix
}
\description{
This function combines multiple PGS into one score
}
