% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_PRS.R
\name{compute_PRS}
\alias{compute_PRS}
\title{Estimate PRS}
\usage{
compute_PRS(geno, weight_file, out, start_col = 4, ispgen = F)
}
\arguments{
\item{geno}{Genotype file in plink format (bed/bim/fam).}

\item{weight_file}{The per-allele SNP effect}

\item{out}{Name of output file.}

\item{start_col}{Index of starting column of SNP effect sizes to estimate PRS in the weight file (DEFAULT = 4).}
}
\value{
Execute plink. Return 0 if successful.
}
\description{
This function estimate PRS with a reference scores
}
