% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PRS_acc.R
\name{eval_single_PRS}
\alias{eval_single_PRS}
\title{Evaluate prediction accuracy of a single PRS}
\usage{
eval_single_PRS(
  data_df,
  pheno = "trait",
  prs_name,
  covar_list,
  isbinary = F,
  liabilityR2 = F,
  alpha = 0.05
)
}
\arguments{
\item{data_df}{Data to assess prediction accuracy}

\item{prs_name}{PGS list of the trait}

\item{covar_list}{Array of covariates}

\item{isbinary}{TRUE if binary and FALSE otherwise (default = FALSE)}

\item{liabilityR2}{TRUE if liability R2 should be reported (default = FALSE)}

\item{alpha}{Significance level to estimate power (default = 0.05)}
}
\value{
A dataframe for prediction accuracy of a single PRS and their power with R2, R2 for output form, standard error, lower 95\% CI, upper 95\% CI, P-value and power
}
\description{
This function get prediction accuracy for a single PRS
}
