% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.1.extract_PGSid.R
\name{extract_PGSid}
\alias{extract_PGSid}
\title{Get PGS id}
\usage{
extract_PGSid(data, trait, efo)
}
\arguments{
\item{data}{directory to metadata score from PGS catalog (pgs_all_metadata_scores.csv)}

\item{trait}{The name of the trait}

\item{efo}{Ontology Trait ID (EFO from PGS catalog)}
}
\value{
a dataframe of PGS id from PGS catalog
}
\description{
This function extract PGS id from EFO code from PGS Catalog
}
