% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varPRS_theory.R
\name{get_acc_prslist_optimized}
\alias{get_acc_prslist_optimized}
\title{Get risk allele by consensus across multiple PRS snp effect panels}
\usage{
get_acc_prslist_optimized(data_df, pgs_list, isbinary = F)
}
\arguments{
\item{data_df}{Data to assess prediction accuracy}

\item{pgs_list}{PGS list of the trait}

\item{isbinary}{TRUE if binary and FALSE otherwise}
}
\value{
A dataframe for prediction accuracy of PRS and their power
}
\description{
This function get the risk-increasing allele based on consensus across PRS panels
}
