% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.1.get_risk_allele.R
\name{get_risk_allele}
\alias{get_risk_allele}
\title{Get risk allele by consensus across multiple PRS snp effect panels}
\usage{
get_risk_allele(
  traitname = "cad",
  pgslist = "~/data/optimization/cad/cad_list.txt",
  pgsmetafile = "~/data/pgs_all_metadata_scores.csv",
  freqfile =
    "/home/jupyter/data/AoU_98K_WGS_QCed_callrate.0.9_hwe.1e-15_maf0.0001_eur.afreq",
  anc = "eur",
  out = "RiskAllele_cad_eur.txt"
)
}
\arguments{
\item{traitname}{Name of the trait}

\item{pgslist}{PGS list of the trait}

\item{pgsmetafile}{Metafile from PGS catalog (pgs_all_metadata_scores.csv)}

\item{freqfile}{Allele frequency file by plink2, ends with .afreq}

\item{anc}{Ancestry}

\item{out}{Output}
}
\value{
A dataframe of SNP and risk-increasing allele
}
\description{
This function get the risk-increasing allele based on consensus across PRS panels
}
