% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_snpeffect_toALT.R
\name{harmonize_snpeffect_toALT}
\alias{harmonize_snpeffect_toALT}
\title{Harmonize snp effects across PGS to the alternative allele}
\usage{
harmonize_snpeffect_toALT(ref_file, pgs_folder, pgs_list, out)
}
\arguments{
\item{ref_file}{Reference file contain SNP ID (ID), reference allele (REF) and alternative allele (ALT) columns (e.g allele frequency output --freq from PLINK2)}

\item{pgs_folder}{Directory to folder contain PGS}

\item{pgs_list}{File contain file names of single PGS on each line. The files must exist in the pgs_folder folder}

\item{out}{Filename of the output}
}
\value{
A file is written containing SNP effects harmonized to the ALT allelle with columns SNP ID, ALT allele and BETAs
}
\description{
This function harmonize SNP effect across PGS to the alternative allele
}
