% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_columns.R
\name{make_dummy_columns}
\alias{make_dummy_columns}
\title{Make dummy columns for categorical variables}
\usage{
make_dummy_columns(filename, cat_covar_list)
}
\arguments{
\item{filename}{File name with variables}

\item{cat_covar_list}{Categorical variables}
}
\value{
This function will generate dummy columns for categorical variables
\itemize{
\item Return data frame with dummy variables of categorical variables
}
}
\description{
This function makes dummy columns for categorical variables
}
